implementação de tela e rest

+++ src/main/java/br/com/docket/teste/service/impl/CartorioServiceImpl.java
@@ -1,6 +1,7 @@
 package br.com.docket.teste.service.impl;
 
 import br.com.docket.teste.domain.Cartorio;
+import br.com.docket.teste.exception.CartorioNaoAchadoException;
 import br.com.docket.teste.repository.CartorioRepository;
 import br.com.docket.teste.service.CartorioService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -39,13 +40,19 @@ public class CartorioServiceImpl implements CartorioService {
 
     @Override
     public void alterar(Integer id, String nome) {
-        Cartorio cartorio = repository.findById(id).get();
+        Optional<Cartorio> optCartorio = repository.findById(id);
+        if (!optCartorio.isPresent()) {
+            throw new CartorioNaoAchadoException();
+        }
+        Cartorio cartorio = optCartorio.get();
         cartorio.setNome(nome);
         repository.save(cartorio);
     }
 
     @Override
     public void excluir(Integer id) {
+        if (!repository.existsById(id))
+            throw new CartorioNaoAchadoException();
         repository.deleteById(id);
     }
